# Conflicts:
#	MyBachelor.pro.user

+++ schedule.cpp
@@ -20,6 +20,10 @@ Schedule::Schedule(QWidget*parent): QWidget(parent)  {
     QObject::connect(api, &BackendlessAPI::itemAdded, this, [&](){
         updateData();
     });
+    QObject::connect(api, &BackendlessAPI::deleteItemFromTableSuccess, this, [&](){
+        updateData();
+    });
+
 
     QObject::connect(api, &BackendlessAPI::loadTableItemsSuccess, this, [&](auto replyValue){
         qDebug() << "Loaded " << replyValue;
@@ -33,8 +37,10 @@ Schedule::Schedule(QWidget*parent): QWidget(parent)  {
         default:
             return;
         }
-
         auto jsonObject = jsonResponse.array();
+
+        calendar->clearContents();
+
         for (const auto& item : jsonObject) {
             auto lessonObject = item.toObject();
 
@@ -51,59 +57,79 @@ Schedule::Schedule(QWidget*parent): QWidget(parent)  {
             calendar->setItem(dayOfWeek, hourStart, someItem);
         }
     });
-    QObject::connect(editMode, &QPushButton::clicked, this, [&](){
+
+    //For debuging
+    QObject::connect(calendar, &QTableWidget::cellClicked, this, [&](){
+        qDebug() << calendar->currentRow() << " Row";
+        qDebug() << calendar->currentColumn() << " Column";
+        // qDebug() << calendar->currentItem()->text();
+    });
+
+    QObject::connect(deleteItemButton, &QPushButton::clicked, this, [&](){
         auto dayOfWeek = calendar->currentRow();
         auto hourStart = calendar->currentColumn();
 
         auto item = calendar->item(dayOfWeek, hourStart);
+        if (!item) {
+            qDebug() << "ITEM IS NOT SELECTED!!!";
+            notDeletable.show();
+            return;
+        }
         auto objectId = item->data(Qt::UserRole);
 
-        qDebug() << objectId;
 
-        api->deleteItemFromTable("Schedules", objectId.toString());
+        qDebug() << "objectId" << objectId;
 
-        //popUpWindow->show();
+        if(item->text() == ""){
+            notDeletable.exec();
+        }else{
+            api->deleteItemFromTable("Schedules", objectId.toString());
+        }
     });
-
-
+    QObject::connect(editMode, &QPushButton::clicked, this, [&](){
+        popUpWindow->show();
+    });
 
     calendar->setVisible(true);
 
     editMode->setText(Schedule::tr("Edit Mode"));
+    deleteItemButton->setText(Schedule::tr("deleteClass"));
+    notDeletable.setText(Schedule::tr("NothingHere"));
+    notDeletable.setInformativeText(Schedule::tr("CanNotDelete"));
+
     dateLay->addSpacing(400);
     dateLay->addWidget(date);
-    dateLay->addSpacing(400);
+    dateLay->addSpacing(385);
     dateLay->addWidget(editMode);
+    dateLay->addWidget(deleteItemButton);
     table->addLayout(dateLay);
     dateLay->addSpacing(350);
     table ->addWidget(calendar);
     table->addStretch();
     setLayout(table);
 
-    calendar->setVerticalHeaderItem(0, monday);
-    calendar->setVerticalHeaderItem(1, tuesday);
-    calendar->setVerticalHeaderItem(2, wednesday);
-    calendar->setVerticalHeaderItem(3, thursday);
-    calendar->setVerticalHeaderItem(4, friday);
+    dayItems.push_back(monday);
+    dayItems.push_back(tuesday);
+    dayItems.push_back(wednesday);
+    dayItems.push_back(thursday);
+    dayItems.push_back(friday);
 
-    calendar->setHorizontalHeaderItem(0, zero);
-    calendar->setHorizontalHeaderItem(1, one);
-    calendar->setHorizontalHeaderItem(2, two);
-    calendar->setHorizontalHeaderItem(3, three);
-    calendar->setHorizontalHeaderItem(4, four);
-    calendar->setHorizontalHeaderItem(5, five);
-    calendar->setHorizontalHeaderItem(6, six);
-    calendar->setHorizontalHeaderItem(7, seven);
-    calendar->setHorizontalHeaderItem(8, eight);
-    calendar->setHorizontalHeaderItem(9, nine);
+    classItems.push_back(zero);
+    classItems.push_back(one);
+    classItems.push_back(two);
+    classItems.push_back(three);
+    classItems.push_back(four);
+    classItems.push_back(five);
+    classItems.push_back(six);
+    classItems.push_back(seven);
+    classItems.push_back(eight);
+    classItems.push_back(nine);
 
+    setupUI();
 
-
-    calendar->setRowHeight(0, 75);
-    calendar->setRowHeight(1, 75);
-    calendar->setRowHeight(2, 75);
-    calendar->setRowHeight(3, 75);
-    calendar->setRowHeight(4, 75);
+    for(int i = 0; i < 5; i++){
+        calendar->setRowHeight(i, 75);
+    }
 
     calendar->setFixedSize(1086,402);
 
@@ -113,6 +139,20 @@ Schedule::Schedule(QWidget*parent): QWidget(parent)  {
     date->setText(ctime(&timestamp));
 
     updateData();
+
+    //qDebug()<< dayItems[0];
+
+}
+
+void Schedule::setupUI() {
+    for(int i = 0; i < 10; i++){
+        if(i < 5){
+            calendar->setVerticalHeaderItem( i, dayItems[i]);
+            calendar->setHorizontalHeaderItem( i, classItems[i]);
+        }else{
+            calendar->setHorizontalHeaderItem( i, classItems[i]);
+        }
+    }
 }
 
 Schedule::~Schedule(){}
